feat(tenant): 应用管理增加真删除和恢复功能
- 在 Tenant 和 TenantApp API 中添加真删除和恢复接口
- 更新前端页面，增加真删除和恢复操作按钮
- 优化应用管理表格，添加状态开关和备注字段
- 修复部分功能的权限控制问题

+++ src/modules/tenant/views/TenantApp/components/GetTableColumns.tsx
@@ -12,7 +12,7 @@ import type { TenantAppVo } from '~/tenant/api/TenantApp.ts'
 import type { UseDialogExpose } from '@/hooks/useDialog.ts'
 
 import { useMessage } from '@/hooks/useMessage.ts'
-import { deleteByIds } from '~/tenant/api/TenantApp.ts'
+import { deleteByIds, realDelete, recovery, save } from '~/tenant/api/TenantApp.ts'
 import { ResultCode } from '@/utils/ResultCode.ts'
 import hasAuth from '@/utils/permission/hasAuth.ts'
 
@@ -32,11 +32,61 @@ export default function getTableColumns(dialog: UseDialogExpose, formRef: any, t
     // 普通列
     { label: () => t('tenant.tenantId'), prop: 'tenant_id' },
     { label: () => t('tenantApp.appName'), prop: 'app_name' },
-    { label: () => t('tenantApp.appKey'), prop: 'app_key' },
-    { label: () => t('tenantApp.appSecret'), prop: 'app_secret' },
-    { label: () => t('tenantApp.status'), prop: 'status' },
+    {
+      label: () => t('tenantApp.appKey'), prop: 'app_key',
+      width: 160,
+      cellRenderTo: {
+        name: 'nmCellEnhance',
+        props: {
+          type: 'copy',
+        },
+      },
+    },
+    {
+      label: () => t('tenantApp.appSecret'), prop: 'app_secret',
+      width: 310,
+      cellRenderTo: {
+        name: 'nmCellEnhance',
+        props: {
+          type: 'copy',
+        },
+      },
+    },
+    {
+      label: () => t('tenantApp.status'), prop: 'status',
+      width: 80,
+      cellRenderTo: {
+        name: 'nmCellEnhance',
+        props: {
+          type: 'switch',
+          prop: 'status',
+          props: {
+            size: 'small',
+            activeValue: true,
+            inactiveValue: false,
+            on: {
+              change: (value: boolean, row: any, proxy: MaProTableExpose) => {
+                console.log('value', value)
+                save(row.id, {
+                  ...row,
+                  status: value,
+                }).then((res) => {
+                  if (res.code === ResultCode.SUCCESS) {
+                    msg.success(t('crud.updateSuccess'))
+                    proxy.refresh()
+                  }
+                  else {
+                    msg.error(t('crud.updateError'))
+                  }
+                })
+              },
+            },
+          },
+        },
+      },
+    },
     { label: () => t('tenantApp.description'), prop: 'description' },
-    { label: () => t('tenantApp.createdAt'), prop: 'created_at' },
+    { label: () => t('tenantApp.createdAt'), prop: 'created_at', width: 180 },
 
     // 操作列
     {
@@ -57,18 +107,44 @@ export default function getTableColumns(dialog: UseDialogExpose, formRef: any, t
             },
           },
           {
-            name: 'del',
-            show: ({ row }) => showBtn('tenantApp:tenant_app:delete', row),
-            icon: 'i-heroicons:trash',
-            text: () => t('crud.delete'),
+            name: 'recovery',
+            icon: 'i-heroicons:arrow-left-start-on-rectangle',
+            show: ({ row }) => showBtn('tenant:tenantApp:recovery', row) && row.deleted_at !== null,
+            text: () => t('crud.restore'),
             onClick: ({ row }, proxy: MaProTableExpose) => {
-              msg.delConfirm(t('crud.delDataMessage')).then(async () => {
-                const response = await deleteByIds([row.id])
+              msg.confirm(t('crud.restoreMessage')).then(async () => {
+                const response = await recovery([row.id])
                 if (response.code === ResultCode.SUCCESS) {
-                  msg.success(t('crud.delSuccess'))
+                  msg.success(t('crud.restoreSuccess'))
                   await proxy.refresh()
                 }
               })
+            },
+          },
+          {
+            name: 'del',
+            show: ({ row }) => showBtn('tenantApp:tenantApp:delete', row),
+            icon: 'i-heroicons:trash',
+            text: () => t('crud.delete'),
+            onClick: ({ row }, proxy: MaProTableExpose) => {
+              if (row?.deleted_at !== null && showBtn('tenant:tenantApp:realDelete', row)) {
+                msg.delConfirm(t('crud.realDeleteDataMessage')).then(async () => {
+                  const response = await realDelete([row.id])
+                  if (response.code === ResultCode.SUCCESS) {
+                    msg.success(t('crud.delSuccess'))
+                    await proxy.refresh()
+                  }
+                })
+              }
+              else {
+                msg.delConfirm(t('crud.delDataMessage')).then(async () => {
+                  const response = await deleteByIds([row.id])
+                  if (response.code === ResultCode.SUCCESS) {
+                    msg.success(t('crud.delSuccess'))
+                    await proxy.refresh()
+                  }
+                })
+              }
             },
           },
         ],
